requirements.txt
crewai>=0.28.8
crewai-tools
python-dotenv
requests
pydantic
streamlit>=1.28.0